+++ docs/changes.rst
@@ -9,6 +9,7 @@ Not yet released.
 * Display ignored checks and allow to make them active again.
 * Display nearby keys on monolingual translations.
 * Added support for groupping string shapings.
+* Recommend upgrade to new Weblate version in the system checks.
 
 Weblate 3.10.3
 --------------